Add license. And refactor... its still broken.
Working on a re-write to communicate back and forth over channels. This
is to learn channels but also let the worker threads fetch the next unit
of work as they finish one unit, rather than the current implementation
(which equally divides the files among the threads). Current
implementation can mean some threads are running a long time while
others finish (if some threads get through an equal number of files
faster than other threads). So it'll be a fun re-write. Concept is
there, but it's currently broken at the moment. Should've checked out a
new branch and done the work there, oops!